chore: move util/sql to pkg/util/sql
